Move home components to own components
