+++ client/components/home/home.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import {HomeCarousel} from './HomeCarousel'
+import {HomeFooter} from './HomeFooter'
+
+export function Home() {
+  return (
+    <div>
+      <HomeCarousel />
+      <HomeFooter />
+    </div>
+  )
+}
