+++ client/components/index.js
@@ -6,7 +6,9 @@
 export {default as Navbarclass} from './navbar'
 export {default as CategoryBar} from './cateogrybar'
 export {default as UserHome} from './user-home'
-export {HomeCarousel} from './HomeCarousel'
+export {Home} from './home/home'
+export {HomeCarousel} from './home/HomeCarousel'
+export {HomeFooter} from './home/HomeFooter'
 export {Login, Signup} from './auth-form'
 export {default as Products} from './products'
 export {default as Product} from './product'
